Rollback to spring version 2.4
For some unknown reason, when I was using 2.6, everything was just a
mess. Swagger springfox would not work, and everytime I tried to start
the server it tried to pick up 'ghost' files like when I moved a class
inside a package, it would pick up the file in the previous location
thus causing bean dependency problems. Maybe 2.6 has breaking changes or
whatever I am not aware of as I have not read it, but for now 2.4 works.